# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cloe API Documentation",
        "description": "Cloe API documentation",
        "contact": {
            "name": "Jesus Romero",
            "email": "gisuss8520@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://cloeBE.test",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/api/users/index": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "INDEX Users",
                "description": "Retorna la lista de usuarios registrados en el sistema",
                "operationId": "73210a5cd9eafccfe1c4fc8a17c614cf",
                "responses": {
                    "200": {
                        "description": "Recurso obtenido con éxito.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/register": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "STORE a new user",
                "description": "Registra un nuevo usuario.",
                "operationId": "b229036599740c94279c740652a3fbf9",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recurso creado con éxito.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/show/{user}": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "SHOW User",
                "description": "Retorna la información del usuario solicitado.",
                "operationId": "c543c7e85391a1480bc8df62dca23c18",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Parámetro necesario para la consulta en la tabla usuarios",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        },
                        "examples": {
                            "integer": {
                                "summary": "Introduce el número de id de un usuario.",
                                "value": 1
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recurso obtenido con éxito.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "404": {
                        "description": "No se ha encontrado el usuario.",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error.",
                        "content": {
                            "application/json": {}
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/getByRoleName": {
            "get": {
                "tags": [
                    "Users"
                ],
                "summary": "Listado de users by role name",
                "description": "Retorna la lista de usuarios dado un rol",
                "operationId": "3f213bb95206fc6f02786257c6fb1466",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recurso obtenido con éxito.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/update/{user}": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "UPDATE user",
                "description": "Actualiza un usuario.",
                "operationId": "c8f3975516cded7db01eca85767e7101",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Parámetro necesario para la búsqueda en la tabla usuarios",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "integer": {
                                "summary": "Introduce el número de id de un usuario.",
                                "value": 1
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recurso actualizado con éxito.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/first-reset-password/{user}": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "First reset password",
                "description": "Restablecimiento de contraseña por primera vez.",
                "operationId": "2950b8741f40e8b00f511c5a9c6e443d",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Parámetro necesario para la búsqueda en la tabla usuarios",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "integer": {
                                "summary": "Introduce el número de id de un usuario.",
                                "value": 1
                            }
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "confirm_password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "401": {
                        "description": "Returns when user is not authenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Not authorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/desactivate/{user}": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "DESACTIVATE an User by ID",
                "description": "Desactiva un usuario del sistema",
                "operationId": "89415432b8a279a496383c2667e5ee0a",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Parámetro necesario para la búsqueda en la tabla usuarios",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "integer": {
                                "summary": "Introduce el número de id de un usuario.",
                                "value": 1
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recurso obtenido con éxito.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/activate/{user}": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "ACTIVATE an User by ID",
                "description": "Activa un usuario del sistema",
                "operationId": "75e09c970c6eb5a3f2ba7bd560d031fb",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Parámetro necesario para la búsqueda en la tabla usuarios",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "integer": {
                                "summary": "Introduce el número de id de un usuario.",
                                "value": 1
                            }
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/activarUserRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Recurso obtenido con éxito.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/users/delete/{user}": {
            "delete": {
                "tags": [
                    "Users"
                ],
                "summary": "DELETE {user}",
                "description": "Elimina un usuario.",
                "operationId": "d4c84189d2dd891d1a94d585a4b1665a",
                "parameters": [
                    {
                        "name": "user",
                        "in": "path",
                        "description": "Parámetro necesario para la búsqueda en la tabla usuarios",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "examples": {
                            "int": {
                                "summary": "Introduce el número de id de un usuario.",
                                "value": "1"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Recurso eliminado con éxito.",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "default": {
                        "description": "Ha ocurrido un error."
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CreateUserRequest": {
                "properties": {
                    "name": {
                        "description": "Nombre",
                        "type": "string",
                        "example": "Juan",
                        "nullable": false
                    },
                    "lastname": {
                        "description": "Apellido",
                        "type": "string",
                        "example": "Arango",
                        "nullable": false
                    },
                    "ci_type": {
                        "description": "Tipo de identificación personal",
                        "type": "string",
                        "example": "V",
                        "nullable": false
                    },
                    "ci_number": {
                        "description": "Número de identificación personal",
                        "type": "string",
                        "example": "20753800",
                        "nullable": false
                    },
                    "email": {
                        "description": "Correo eletrónico",
                        "type": "string",
                        "format": "email",
                        "nullable": false
                    },
                    "estado_id": {
                        "description": "ID de estado",
                        "type": "integer",
                        "example": "7",
                        "nullable": false
                    },
                    "municipio_id": {
                        "description": "ID de municipio",
                        "type": "integer",
                        "example": "86",
                        "nullable": false
                    },
                    "address": {
                        "description": "Dirección del usuario",
                        "type": "string",
                        "example": "Valencia, calle 101, Venezuela",
                        "nullable": false
                    },
                    "role": {
                        "description": "Rol",
                        "type": "string",
                        "example": "Recolector",
                        "nullable": false
                    }
                },
                "type": "object"
            },
            "RoleRequest": {
                "properties": {
                    "roleName": {
                        "description": "Nombre de rol",
                        "type": "string",
                        "example": "Recolector",
                        "nullable": false
                    },
                    "estado_id": {
                        "description": "ID de estado",
                        "type": "integer",
                        "example": "7",
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
            "activarUserRequest": {
                "properties": {
                    "role": {
                        "description": "Rol",
                        "type": "string",
                        "example": "Recolector",
                        "nullable": false
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "name": "sanctum",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}